Data driven UI

+++ plugins/cc-cli/src/cli/save-data-template.ts
@@ -1,10 +1,16 @@
 import path from 'path';
 import fs from 'fs';
-import { dynamicRequire } from '@component-controls/core/node-utils';
+import deepEqual from 'fast-deep-equal';
+import {
+  dynamicRequire,
+  relativeImport,
+} from '@component-controls/core/node-utils';
+import { getDataFile } from '@component-controls/instrument';
 import { log } from '@component-controls/logger';
 import { CliOptions, getTestFolder } from './utils';
-import { TemplateOptions, DataImportOptions, relativeImport } from '../utils';
+import { TemplateOptions, DataImportOptions } from '../utils';
 import { createDataTemplate } from '../data-templates/data-template';
+import { BuildConfiguration } from '@component-controls/core';
 
 /**
  * save a data template file based on options
@@ -12,16 +18,14 @@ import { createDataTemplate } from '../data-templates/data-template';
  */
 export const saveDataTemplate = async <P extends TemplateOptions>(
   options: CliOptions<P>,
+  configuration?: BuildConfiguration,
 ): Promise<DataImportOptions | undefined> => {
   const testFolder = getTestFolder(options);
   if (!testFolder) {
     return undefined;
   }
   const { test, overwrite } = options;
-  const dataName = test
-    .split('.')
-    .map((e, i) => (i === 1 ? 'data' : e))
-    .join('.');
+  const dataName = getDataFile(test);
 
   const filePath = path.resolve(testFolder, dataName);
 
@@ -35,13 +39,21 @@ export const saveDataTemplate = async <P extends TemplateOptions>(
       return undefined;
     }
   }
-  const dataTemplate = await createDataTemplate(options, existing);
+  const dataTemplate = await createDataTemplate(
+    options,
+    existing,
+    configuration,
+  );
   if (dataTemplate) {
-    log('saving data', filePath, [184, 226, 255]);
-    if (!fs.existsSync(testFolder)) {
-      fs.mkdirSync(testFolder);
+    const isModified =
+      typeof existing === 'undefined' || deepEqual(existing, dataTemplate.data);
+    if (isModified && overwrite) {
+      log('saving data', filePath, [184, 226, 255]);
+      if (!fs.existsSync(testFolder)) {
+        fs.mkdirSync(testFolder);
+      }
+      fs.writeFileSync(filePath, dataTemplate.content, 'utf8');
     }
-    fs.writeFileSync(filePath, dataTemplate.content, 'utf8');
     return {
       data: dataTemplate.data,
       filePath: relativeImport(testFolder, filePath),
